Fix main merge

+++ src/main.rs
@@ -2,10 +2,10 @@ use nes::cpu::CPU;
 use nes::bus::BUS;
 use rand::Rng;
 use sdl2::event::Event;
+use sdl2::EventPump;
 use sdl2::keyboard::Keycode;
 use sdl2::pixels::Color;
 use sdl2::pixels::PixelFormatEnum;
-use sdl2::EventPump;
 
 fn color(byte: u8) -> Color {
     match byte {
@@ -56,7 +56,7 @@ fn handle_user_input(cpu: &mut CPU, event_pump: &mut EventPump) {
             Event::KeyDown { keycode: Some(Keycode::D), .. } => {
                 cpu.memory_write(0xff, 0x64);
             }
-            _ => { /* do nothing */ }
+            _ => {/* do nothing */}
         }
     }
 }
@@ -68,8 +68,7 @@ fn main() {
     let window = video_subsystem
         .window("Snake game", (32.0 * 10.0) as u32, (32.0 * 10.0) as u32)
         .position_centered()
-        .build()
-        .unwrap();
+        .build().unwrap();
 
     let mut canvas = window.into_canvas().present_vsync().build().unwrap();
     let mut event_pump = sdl_context.event_pump().unwrap();
@@ -77,8 +76,8 @@ fn main() {
 
     let creator = canvas.texture_creator();
     let mut texture = creator
-        .create_texture_target(PixelFormatEnum::RGB24, 32, 32)
-        .unwrap();
+        .create_texture_target(PixelFormatEnum::RGB24, 32, 32).unwrap();
+    
 
     let game_code = vec![
         0x20, 0x06, 0x06, 0x20, 0x38, 0x06, 0x20, 0x0d, 0x06, 0x20, 0x2a, 0x06, 0x60, 0xa9, 0x02,
@@ -104,13 +103,14 @@ fn main() {
         0x60, 0xa6, 0xff, 0xea, 0xea, 0xca, 0xd0, 0xfb, 0x60,
     ];
 
+
     //load the game
     let bus = BUS::new();
     let mut cpu = CPU::new(bus);
     cpu.load(game_code);
     cpu.reset();
 
-    let mut screen_state = [0_u8; 32 * 3 * 32];
+    let mut screen_state = [0 as u8; 32 * 3 * 32];
     let mut rng = rand::thread_rng();
 
     // run the game cycle
@@ -129,4 +129,5 @@ fn main() {
 
         ::std::thread::sleep(std::time::Duration::new(0, 70_000));
     });
+
 }
